jpms: Refactor handlebars.humanize
